basic norm
