+++ main/srcs/objects/object.c
@@ -41,26 +41,3 @@ void    clean_object_from_list(void *obj)
     object->clean(object->object);
     free(object);
 }
-
-void    *get_object_from_list(t_list *lst, int type)
-{
-    t_object    *object;
-
-    while (lst != NULL)
-    {
-        object = lst->content;
-        if (object->type == type)
-            return (object);
-        lst = lst->next;
-    }
-    return (NULL);
-}
-
-int   none_test_intersection(void *obj, t_interparam *param)
-{
-    (void)obj;
-    (void)param;
-
-    printf("this oject has no test intersection function\n");
-    exit(ERROR);
-}
