+++ main/srcs/objects/plane.c
@@ -1,6 +1,5 @@
 #include "../../inc/minirt.h"
 
-double pl_cale_dist(t_ray *ray, t_plane *pl, int *hit, t_interparam *param);
 void    print_plane(void    *pl)
 {
     t_plane *plane;
@@ -22,26 +21,22 @@ void    clean_plane(void *pl)
     free(plane);
 }
 
-
 int     pl_test_intersection(void *object, t_interparam *param)
 {
     t_plane     *plane;
     t_object    *obj;
 
-
     obj = (t_object *)object;
     plane = (t_plane *)obj->object;
-
     return (pl_sub_intersect(plane, param));
-
 }
 
 int pl_sub_intersect(t_plane *plane, t_interparam *param)
 {
     double      dist;
     t_vec3d     *ray;
-    ray = &param->ray->direction;
 
+    ray = &param->ray->direction;
     dist = pl_cale_dist(param->ray, plane, &param->f_ishit, param);
     if (dist > 0)
     {
@@ -54,36 +49,4 @@ int pl_sub_intersect(t_plane *plane, t_interparam *param)
         param->f_ishit = 1;
     }
     return (param->f_ishit);
-
 }
-
-// PROB
-// plane [0,0,0] , normal = [0,0,1]
-double pl_cale_dist(t_ray *ray, t_plane *pl, int *hit, t_interparam *param)
-{
-    double dist;
-    double dot;
-    t_vec3d     inters_point;
-
-    dist = -1;
-    dot = vec3d_dot(&pl->normal, &ray->direction);
-    if (almost_equal(dot, 0))
-        return (0);
-    vec3d_minus(&inters_point, &pl->point, &ray->origin);
-    dist = vec3d_dot(&inters_point, &pl->normal) / dot;
-    if (dist < 0)
-        return (0);
-    ft_memcpy(&param->f_normal, &pl->normal, sizeof(t_vec3d));
-    if (dot > 0)
-    {
-        param->f_normal.x = -param->f_normal.x;
-        param->f_normal.y = -param->f_normal.y;
-        param->f_normal.z = -param->f_normal.z;
-    }
-    return (dist);
-}
-
-int almost_equal(double a, double b)
-{
-    return (fabs(a - b) < EPSILON);
-}
