+++ main/srcs/utils/mlx_util.c
@@ -1,11 +1,5 @@
 #include "../../inc/minirt.h"
 
-static int	keyhandler(int keycode);
-/// @brief mlx_my_putpixel is a function that put a pixel in the image.
-/// @param img_data
-/// @param width line length
-/// @param bpp bits per pixel
-/// @param color 
 void	mlx_my_putpixel(t_img *data, int x, int y, int color)
 {
 	char	*dst;
@@ -14,7 +8,6 @@ void	mlx_my_putpixel(t_img *data, int x, int y, int color)
 	*(unsigned int*)dst = color;
 }
 
-/// @brief init mlx window and image
 void    mlx_my_init(t_mlx_confix *vars)
 {
     vars->mlx = mlx_init();
@@ -25,7 +18,6 @@ void    mlx_my_init(t_mlx_confix *vars)
     vars->img.addr = mlx_get_data_addr(vars->img.img, &vars->img.bpp, &vars->img.width, &vars->img.endian);
     fill_image_with_color(vars);
     vars->is_init_img = 1;
-
 }
 
 void fill_image_with_color(t_mlx_confix *vars)
@@ -48,86 +40,22 @@ void fill_image_with_color(t_mlx_confix *vars)
     }
 }
 
-/// @brief function handle close window
-// int    mlx_my_close(int key, t_mlx_confix *vars)
-// {
-//     (void)key;
-//     if (vars->is_init_img)
-//         mlx_destroy_image(vars->mlx, vars->img.img);
-//     if (vars->is_initwin)
-//         mlx_destroy_window(vars->mlx, vars->win);
-//     return (0);
-// }
-
-static int	close_window(void)
+int close_window(void)
 {
-	exit (0);
+    exit(0);
 }
 
-// TODO: remove this function
-int animation(void *program)
-{
-    t_prog *prog;
-	int		win_x;
-	int		win_y;
-	int		color;
-    int     loop;
-    int     r;
-    int     g;
-    int     b;
-    t_mlx_confix *vars;
-
-    prog = (t_prog *)program;
-    vars = &(prog->mlx_config);
-    loop = 0;
-    while (loop < 20000)
-    {
-        win_y = 0;
-        while (win_y < WINDOW_HEIGHT)
-        {
-	        win_x = 0;
-            while (win_x < WINDOW_WIDTH)
-            {
-                // all render logic here
-                r = ((double)win_x / (double)WINDOW_WIDTH) * 255.0;
-                g = ((double)win_y / (double)WINDOW_HEIGHT) * 255.0;
-                b = (((double)win_y *  (double)win_x))  / ((double)WINDOW_HEIGHT * (double)WINDOW_WIDTH) * 255.0;
-
-                color = color_rgb2int(255, 0, 0);
-                if (win_x == win_y)
-                    mlx_my_putpixel(&(prog->mlx_config.img), win_x, win_y, color);
-                // mlx_my_putpixel(&(prog->mlx_config.img), WINDOW_WIDTH - win_x , WINDOW_HEIGHT - win_y, color * loop);
-                win_x++;
-            }
-            mlx_put_image_to_window(vars->mlx, vars->win, vars->img.img, 0, 0);
-            usleep(10);
-            win_y++;
-        }
-        loop++;
-        }
-}
-
-/// @brief function handle loop
 void prog_mlx_loop(t_prog *prog)
 {
     t_mlx_confix *vars;
 
     vars = &(prog->mlx_config);
-    // TODO Bug When exist
-    // mlx_hook(vars->win, 2, 0, mlx_my_close, vars);
-    // X red
     mlx_hook(vars->win, 17, 0, close_window, vars);
-
     mlx_hook(vars->win, 2, 1L << 0, &keyhandler, vars);
-
-    // Simple animation
-    // mlx_loop_hook(vars->mlx, animation, prog);
-
-    // SWITCH turn on/off mlx window
     mlx_loop(vars->mlx);
 }
 
-static int	keyhandler(int keycode)
+int	keyhandler(int keycode)
 {
 	if (keycode == KEY_LEFT || keycode == KEY_A)
 		printf("KEY_A: %s\n", "LEFT");
@@ -177,8 +105,6 @@ static int	keyhandler(int keycode)
 		printf("KEY_ESC: %s\n", "EXIT");
 	return (0);
 }
-
-
 
 void debug_mlx_status(t_mlx_confix *vars)
 {
