+++ main/srcs/utils/my_math.c
@@ -1,6 +1,5 @@
 #include "../../inc/minirt.h"
 
-// return min exist positive value
 double solve_quadratic(double a, double b, double c)
 {
     double test;
@@ -25,4 +24,3 @@ double solve_quadratic(double a, double b, double c)
     }
     return (-1);
 }
-
